Rewrite the rewrite
Russel managed to nerd snipe me pretty good with his comments on the PR
so I've spent a lot of time trying to figure out a more parallel method
for managing the LRU.

This approach changes things so that every entry in the cache has an
associated pid that manages all of the ets table entries. This way we
can insert and delete all of our LRU timestamps in parallel.

Initial testing shows throughput in the millions of operations per
second.

Also for an aside, I don't think the original suggested appraoch for
having individual clients write to ets tables is going to work as well.
It worries me that we'll end up with some strange combination of
operations that ends up orphaning entries. But it seems like the
parallelism of the maximum cache size should be fairly acceptable.